Add support for XOAUTH2 and OAUTHBEARER authentication
This patch adds support for RFC 7628 OAUTHBEARER style authentication
to the nginx mail proxy module for IMAP, POP and SMTP. To help with
legacy libraries, it also implements the pre-RFC XOAUTH2 protocol as well.

It adds ngx_*_auth_{oauthbearer,xoauth2} states, constants, handlers, etc.
The bearer token provided by the client is passed to the backend auth
server in the `Auth-Pass` header.

The auth server may return an additional optional response
header `Auth-Error-Sasl`. It's expected in the auth failure
case that the backend auth server will generate a base64
encoded JSON object that conforms to the error reporting in
https://datatracker.ietf.org/doc/html/rfc7628#section-3.2.2 in this
header.

If present, the value in this header is prefixed with a `+ ` (IMAP) or
`334 ` (SMTP or POP) and returned as the SASL response. We then wait for
a valid line from the client (which we ignore) and then we exit the SASL
mode and return back to standard protocol parsing. If we don't receive
a valid line, we terminate the connection. There's an example of this
looks like in https://datatracker.ietf.org/doc/html/rfc7628#section-4.3